mueve datos a data.js para luego importar en las funciones necesarias
